% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADePTR.R
\docType{package}
\encoding{UTF-8}
\name{ADePTR}
\alias{ADePTR}
\alias{ADePTR-package}
\title{ADePTR: Acoustic Detection Processing and Visualization Tool in R}
\description{
ADePTR is package developed to streamline processing and visualization
of animal movement data from a passive acoustic array.
}
\details{
This package was designed with the following workflow in mind:

\enumerate{
  \item Process input files
  \item Filter erroneous detections
  \item Calculate centers of activity
  \item Summarize detections on a grid of the study area
  \item Interpolate locations and animate
}

For each step of the workflow, there are both possible inputs and
possible outputs. The package is designed to work well with minimal
user input (\emph{i.e.}, most arguments have defaults), but it is also
designed to allow a high level of customization. Advanced users should
be able to edit the outputs of any step before passsing it to the next
step, or alternatively, to enter the workflow with their own data at
any point.
}
\section{Step 1. Process Inputs}{


The primary functions available for this step are:
\itemize{
  \item \code{\link{proc_dets}()} -- uses station/receiver information to
  georeference the detections
  \item \code{\link{plot_sta_history}()} -- makes a summary plot of
  detections over time by station
  \item \code{\link{map_dets}()} -- makes a map showing the number of
  detections by station
}
}

\section{Step 2. Filter}{


The primary functions available for this step are:
\itemize{
  \item \code{\link{spd_filter}()} -- filters detections that would imply
  an unrealistic swim speed
  \item \code{\link{singleton_filter}()} -- filters single detections of
  an individual ID at an individual station over a given time period
}
}

\section{Step 3. COAs}{


The primary functions available for this step are:
\itemize{
  \item \code{\link{coa_locs}()} -- calculates centers of activity (COAs)
  using either arithmetic or harmonic means
  \item \code{\link{map_coas}()} -- makes a map showing the COAs on top of
  detections by station
}
}

\section{Step 4. Summarize on Grid}{


The primary functions available for this step are:
\itemize{
  \item \code{\link{init_raster}()} -- initializes a template raster for
  future functions
  \item \code{\link{rasterize_dets}()} -- converts detection data to counts
  of detections by raster cells
  \item \code{\link{raster_pres}()} -- converts a detections raster to a
  presence-only raster
  \item \code{\link{raster_overlap}()} -- sums the input presence-only or
  presence-absence layers to show overlapping individuals or species
}
}

\section{Step 5. Interpolate and Animate}{


The primary functions available for this step are:
\itemize{
  \item \code{\link{str_paths}()} -- connects locations with straight paths
  \item \code{\link{lc_paths}()} -- connects locations with least
  cost paths
  \item \code{\link{regular_points}()} -- interpolates regular points along
  a straight path or least cost path
  \item \code{\link{animate_points}()} -- creates an animation from regular
  points
}
}

\section{More details}{


For more details on the usage of this package, see the package vignette.

The package has a GitHub repository at
\url{https://github.com/bjsmith-usgs/ADePTR}.
}

