% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{animate_points}
\alias{animate_points}
\title{Simple animation of regular points}
\usage{
animate_points(reg_points, ani.width = 800, ani.height = 600,
...)
}
\arguments{
\item{reg_points}{An object of class \code{reg_points} to create
the animation.}

\item{ani.width}{Width of the animation in pixels. Passed to
\code{\link[animation:ani.options]{animation::ani.options}()}.}

\item{ani.height}{Height of the animation in pixels. Passed to
\code{\link[animation:ani.options]{animation::ani.options}()}.}
}
\value{
Returns an object of class \code{gganim} from package
\link[gganimate:gganimate-package]{gganimate}.
}
\description{
Provides a simple animation of a \code{reg_points} object.
}
\details{
This function quickly creates a simple animation of a
\code{reg_points} object returned by \code{\link{regular_points}()}.
It uses the package \link[gganimate:gganimate-package]{gganimate}
to make an animation. The function returns an object of class
\code{gganim}, so the user can manipulate the object prior to printing
just as they would manipulate any ggplot with \code{`+`}. Printing the
object automatically renders the animation using
\code{\link[gganimate]{animate}}.

We encourage the user to explore
\link[gganimate:gganimate-package]{gganimate} for producing their own
custom animations from the interpolated regular points returned by
\code{\link{regular_points}()}. The source code from this function
could serve as a starting point.

\preformatted{
    ani <- ggplot(reg_points) +
      geom_sf(aes(color = id), key_glyph = "point") +
      theme_bw() +
      transition_time(time = dt) +
      labs(title = 'Time: {frame_time}', color = "ID")
}
}
