% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_proc_inputs.R
\name{as.dets}
\alias{as.dets}
\title{Coerce a \code{data.frame} to class \code{dets}}
\usage{
as.dets(x, crs = 4326)
}
\arguments{
\item{x}{A \code{data.frame} to coerce}

\item{crs}{Coordinate Reference System to use for the detections. Passed
to \code{\link[sf:st_crs]{sf::st_crs}()} to set CRS for sf object.
Defaults to \code{4326}, longitude/latitude on the WGS84 spheroid.}
}
\description{
Coerces a \code{data.frame} to a \code{dets} object
}
