% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_coas.R
\name{coa_locs}
\alias{coa_locs}
\title{Calculate Centers of Activity}
\usage{
coa_locs(proc_det, Delta_t = "1 hour", mean_type = c("arithmetic",
 "harmonic"), ...)
}
\arguments{
\item{proc_det}{A \code{data.frame} of georeferenced detections as returned
by the function \code{\link{proc_dets}()}.}

\item{Delta_t}{The desired time interval for the COAs (\eqn{\Delta}t).}

\item{mean_type}{The type of mean used to calculate the mean positions.
Should be either \code{"arithmetic"} or \code{"harmonic"}.}

\item{...}{Additional arguments (not currently implemented)}
}
\value{
Returns a \code{data.frame} of georeferenced centers of activity.
}
\description{
This function calculates Centers of Activity (COAs) by averaging
the locations of detections in a user-defined time period.
}
\details{
This function implements the two mean position algorithms
suggested by Simpfendorfer et al. (2002) for calculating centers of
activity (COAs).
}
\references{
Simpfendorfer, C.A., M.R. Heupel, and R.E. Hueter. 2002. Estimation of
short-term centers of activity from an array of omnidirectional hydrophones
and its use in studying animal movements. \emph{Canadian Journal of
Fisheries and Aquatic Sciences} 59(1): 23-32.
}
\seealso{
\code{\link{proc_dets}} for details on the formatting of the
\code{data.frame} \code{proc_det}

See Simpfendorfer \emph{et al.} (\href{https://doi.org/10.1139/f01-191}{2002}) for details on using COAs to study
marine animal movements.
}
