% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{filter_line0}
\alias{filter_line0}
\title{Remove length-0 lines from an \code{lcp_ind} object}
\usage{
filter_line0(lcp_ind)
}
\arguments{
\item{lcp_ind}{An object of class \code{lcp_ind}.}
}
\value{
Returns an object of class \code{lcp_ind}.
}
\description{
Takes an object of class \code{lcp_ind} and filters out any
lines that do not have at least 2 vertices.
}
\details{
This function is intended for internal use on \code{lcp_list}
objects.
}
