% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_coas.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Calculate Harmonic Mean}
\usage{
harmonic_mean(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
Returns a vector of length 1
}
\description{
This function is used internally by \code{\link{coa_locs}()} to
calculate the harmonic mean of a set of detections.
}
\details{
Calculates the harmonic mean
(\href{https://en.wikipedia.org/wiki/Harmonic_mean}{Wikipedia}), which
is the reciprocal of the arithmetic mean of the reciprocals, i.e.:

\code{1/(mean(1/x))}
}
\seealso{
Used internally by \code{\link{coa_locs}}
}
