% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_grids.R
\name{init_raster}
\alias{init_raster}
\title{Creates an empty raster for detection summaries}
\usage{
init_raster(xlim = NULL, ylim = NULL, sta = NULL,
                   study_area = NULL, res = 0.01,
                   crs = sp::CRS("+init=epsg:4326"))
}
\arguments{
\item{xlim}{A numeric vector of length 2 giving user-defined x-limits. If
\code{xlim} is provided, \code{ylim} must also be provided. These fields
override both \code{sta} and \code{study_area}.}

\item{ylim}{A numeric vector of length 2 giving user-defined y-limits. If
\code{xlim} is provided, \code{ylim} must also be provided. These fields
override both \code{sta} and \code{study_area}.}

\item{sta}{A data.frame with x- and y-coordinates of all the stations
in the study area. The x- and y-coordinates should be in columns named
\code{"x"} and \code{"y"}, respectively. Ignored if \code{xlim} or
\code{ylim} are provided. This field overrides \code{study_area}.}

\item{study_area}{A polygon of class \code{\link[sf:sfc]{sfc}}
from the package \code{sf} that delimits the study area.}

\item{res}{A numeric vector  of length 1 or 2 giving the size of a grid
cell in units of the coordinate reference system. See
\code{\link[raster]{res}} for details.}

\item{crs}{A character or object of class CRS. Defines the coordinate
reference system to use for the resulting raster. Should match the CRS of
the input if applicable. Used by \code{\link[raster]{raster}} as
the \code{crs} argument. Defaults to \code{(CRS("+init=epsg:4326"))},
indicating EPSG code 4326 for WGS 84 lat/long.}
}
\value{
Returns an object of class \code{RasterLayer}
}
\description{
This function can take one of user-defined x- and y-limits, station
locations, or a study area polygon and create an empty raster to use
in the detection summary figures.
}
\details{
Behavior depends upon the input. Once the function reaches a valid
input, it returns the raster without bothering to check if additional inputs
were provided.

The order it checks in is:
\enumerate{
  \item{\code{xlim} and \code{ylim}} (must provide both)
  \item{\code{sta}} (x- and y-coordinates should be named \code{"x"} and
  \code{"y"})
  \item{\code{study_area}} (polygon of class \code{\link[sf:sfc]{sfc}})
}
}
\examples{
#Load sample data
data(acoustic)

#Using xlim and ylim
xlim <- c(-64.628, -64.612)
ylim <- c(17.770, 17.795)
r.lim <- init_raster(xlim = xlim, ylim = ylim)

#Use UTM
xlim <- c(321160, 335010)
ylim <- c(1965700, 1973810)
r.lim.utm <- init_raster(xlim = xlim, ylim = ylim,
                         res = 500, #units here are meters
                         crs = sp::CRS("+init=epsg:32620"))

#Using sta
r.sta <- init_raster(sta = acoustic$stations)

#Using study_area
r.sa <- init_raster(study_area = acoustic$study_area)

#Change the resolution (in CRS units)
r.sa2 <- init_raster(study_area = acoustic$study_area,
                     res = 0.0001)


}
