% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{is.paths}
\alias{is.paths}
\title{Check if object is of class \code{*_paths}.}
\usage{
is.paths(x)
}
\description{
Convenience function that checks is object is of class \code{*_paths}.
Currently must be either \code{str_paths} or \code{lc_paths}.
}
