% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{lc_path_ind}
\alias{lc_path_ind}
\title{Create least cost paths between each location}
\usage{
lc_path_ind(det_locs_ind, tl)
}
\arguments{
\item{det_locs_ind}{A \code{data.frame} of georeferenced acoustic
detections, subset for a single individual. Could be, \emph{e.g.},
the \code{data.frame} returned by \code{\link{proc_dets}()} or the
\code{data.frame} returned by \code{\link{coa_locs}()}, subset to a
single individual.}

\item{tl}{A \code{\link[gdistance:transition]{TransitionLayer}}
object used to calculate least cost path.}
}
\value{
Returns an boject of class \code{lcp_ind}, which is also
a \code{tbl_df}, \code{tbl}, and \code{data.frame}.
}
\description{
Creates least cost paths as line objects between each location for
a single individual.
}
\details{
This function is intended to be called internally by the
function \code{\link{lc_path}()}.
}
