% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{lc_paths}
\alias{lc_paths}
\title{Create least cost paths between each location}
\usage{
lc_paths(det_locs, conductance)
}
\arguments{
\item{det_locs}{A \code{data.frame} of georeferenced acoustic
detections. Could be, \emph{e.g.}, the \code{data.frame} returned by
\code{\link{proc_dets}()} or the \code{data.frame} returned by
\code{\link{coa_locs}()}.}

\item{conductance}{A \code{\link[raster:Raster-class]{RasterLayer}}
object of conductance, either returned by \code{\link{raster_conduct}()}
or a similar object created by the user.}
}
\value{
Returns an object of class \code{lcp_list}, which is also
a \code{list}. List elements are of class \code{lcp_ind}. See
\code{\link{lc_path_ind}} for details.
}
\description{
Creates least cost paths as line objects between each location for
any number of individuals.
}
\details{
Estimating the least cost paths between a large number of
locations can be time consuming.
}
\examples{

#Load sample data
data(acoustic)

#Process detections and stations
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

#Calculate  COAs -- Note, using 6 hour COAs to reduce processing time
coas.6h <- coa_locs(proc.det, Delta_t = "6 hours")

#Initialize raster
r.cond <- init_raster(study_area = acoustic$study_area, res = 0.001)

#Create conductance raster
conduct <- raster_conduct(sa_rast = r.cond, impassable = acoustic$land)

#Create least-cost paths
lcps <- lc_paths(det_locs = coas.6h, conductance = conduct)

}
