% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_coas.R
\name{map_coas}
\alias{map_coas}
\title{Map Centers of Activity}
\usage{
map_coas(proc_det, coas, coa_crs = 4326, coa_palette = viridis::viridis,
coa_leg_pos = "bottomleft", det_leg_pos = NA, set_par = TRUE, ...)
}
\arguments{
\item{proc_det}{The \code{data.frame} of processed detections used by
\code{\link{coa_locs}()} to create the COAs.}

\item{coas}{The \code{data.frame} of centers of activiy (COAs) created
by \code{\link{coa_locs}()}.}

\item{coa_crs}{A coordinate reference system to use for the COA
locations. Used by \code{\link[sf]{st_sf}} as the \code{crs} argument.
Defaults to integer 4326, indicating EPSG code 4326 for WGS 84 lat/long.}

\item{coa_palette}{A function name that will generate a palette for each
of the individuals tracked. Can be any function that will accept \code{n}
as an argument.}

\item{coa_leg_pos}{Indicates where the COA legend should be (showing
the symbol for each individual animal). Can be any \code{x}
that the base \code{graphics} function \code{\link[graphics]{legend}}
will accept. Defaults to the text string \code{"bottomleft"}.}

\item{det_leg_pos}{Indicates where the default legend for the \emph{
station detections} should be. Passed to \code{\link{map_dets}()} to
control that legend placement. Can be any \code{x} that the base
\code{graphics} function \code{\link[graphics]{legend}} will accept.
Defaults to \code{NA} which prevents that legend from being drawn. Can
be changed to a location \code{!= coa_leg_pos}.}

\item{set_par}{\code{TRUE} or \code{FALSE}. Should the function change
the graphical parameters or not? This should be \code{FALSE} if the
user wishes to manually set the graphical parameters, \emph{e.g.}, so
that they can still add to the plot manually after it is generated.}

\item{...}{Additional arguments to be passed to \code{\link{map_dets}},
\emph{e.g.}, \code{base.layers}, \code{xlim}, \code{ylim}, \emph{etc}.}
}
\description{
Plots a map using \code{\link{map_dets}()} and adds the
COAs returned by \code{\link{coa_locs}()} to the map.
}
\details{
Calls \code{\link{map_dets}()} to draw the basemap and then adds
the COAs on top.
}
\seealso{
\code{\link{map_dets}}
}
