% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_proc_inputs.R
\name{map_dets}
\alias{map_dets}
\title{Plots a map of station locations}
\usage{
map_dets(proc_det, sta.crs = 4326, base.layers = NULL,
             base.borders = NULL, base.cols = NULL,
             sta.col = "black", sta.bg = "red",
             leg.pos = "bottomleft", set.par = TRUE, ...)
}
\arguments{
\item{proc_det}{A \code{data.frame} of class \code{dets} as returned
by the function \code{\link{proc_dets}()}.}

\item{sta.crs}{A coordinate reference system to use for the station
locations. Used by \code{\link[sf]{st_sf}} as the \code{crs} argument.
Defaults to integer 4326, indicating EPSG code 4326 for WGS 84 lat/long.}

\item{base.layers}{A \code{list} of georeferenced vectors to be plotted
as the basemap under the stations. Designed to be \code{\link[sf]{sfc}}
\code{LINESTRING} or \code{POLYGON} objects.}

\item{base.borders}{A \code{list} of colors to used to outline the
polygons listed in \code{base.layers}. Should be the same length as
\code{base.layers}.}

\item{base.cols}{A \code{list} of colors to used to fill the polygons
listed in \code{base.layers}.  Should be the same length as \code{
base.layers}.}

\item{leg.pos}{Indicates where the legend should be. Can be any \code{x}
that the base \code{graphics} function \code{\link[graphics]{legend}}
will accept. Defaults to the text string \code{"bottomleft"}.}

\item{set.par}{\code{TRUE} or \code{FALSE}. Should the function change
the graphical parameters or not? This should be \code{FALSE} if the
user wishes to manually set the graphical parameters, \emph{e.g.}, so
that they can still add to the plot manually after it is generated.}

\item{return_df}{\code{TRUE} or \code{FALSE}. Should the function also
return a \code{data.frame} of aggregated station detections?}

\item{...}{Additional arguments (not currently implemented)}
}
\value{
Plots a figure of station activity over time.
}
\description{
This function takes a processed detection history returned by
\code{\link{proc_dets}()} and plots a map showing the relative number
of detections per station.
}
\details{
Details here
}
\examples{
#Load the example data set
data(acoustic)

#Process detections
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

#Map detections with all defaults
map_dets(proc.det)

#Map detections with custom settings
map_dets(proc_det = proc.det, sta.crs = 4326,
    base.layers = list(acoustic$study_area, acoustic$land),
    base.cols = list("gray30", "wheat"), base.borders = list(NA, "seagreen"),
    xlim=c(-64.628, -64.612), ylim=c(17.770, 17.795),
    leg.pos = "topleft", sta.col = "blue", sta.bg = "orange")

}
\seealso{
\code{\link{proc_dets}} for details on the formatting of the
\code{data.frame} \code{proc_det}
}
