% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_proc_inputs.R
\name{plot.dets}
\alias{plot.dets}
\title{Plot a \code{dets} object}
\usage{
\method{plot}{dets}(proc_det, which = "map", ...)
}
\arguments{
\item{proc_det}{A \code{data.frame} of class \code{dets} as returned
by the function \code{\link{proc_dets}()}.}

\item{which}{A character string for which plot to create. Defaults
to \code{map}. See details below.}

\item{...}{Additional parameters to pass to the plotting function.}
}
\description{
S3 method for plotting a \code{dets} object.
}
\details{
Wrapper that calls one of the plotting functions, depending on the value
of parameter \code{which}. If \code{which == "history"} then the function
\code{\link{plot_sta_history}()} is called (the default), but if
\code{which == "map"} then the function \code{\link{map_dets}()} is called.
}
\examples{
#Load the example data set
data(acoustic)

#Process detections
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

#Map detections
plot(proc.det)

#Plot station history
plot(proc.det, which = "history")
}
