% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{plot.lcp_list}
\alias{plot.lcp_list}
\title{Plot an object of class \code{lcp_list}}
\usage{
\method{plot}{lcp_list}(lcps, ...)
}
\description{
Plots the least cost paths between locations for all individuals
in an \code{lcp_list} object.
}
\details{
See \code{\link{plot_lc_path}()} for details.
}
\examples{

#Load sample data
data(acoustic)

#Process detections and stations
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

#Calculate  COAs -- Note, using 6 hour COAs to reduce processing time
coas.6h <- coa_locs(proc.det, Delta_t = "6 hours")

#Initialize raster
r.cond <- init_raster(study_area = acoustic$study_area, res = 0.001)

#Create conductance raster
conduct <- raster_conduct(sa_rast = r.cond, impassable = acoustic$land)

#Create least-cost paths
lcps <- lc_path(det_locs = coas.6h, conductance = conduct)

#Generic plot
plot(lcps)

#Generic method with land added
plot(lcps, set_par = FALSE)
plot(acoustic$land, add = TRUE, col = "wheat")

}
