% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{plot_lc_path}
\alias{plot_lc_path}
\title{Plot an object of class \code{lcp_list}}
\usage{
plot_lc_path(...)
}
\arguments{
\item{lcps}{An object of class \code{lcp_list}}

\item{path_palette}{A function name that will generate a palette for each
of the individuals tracked. Can be any function that will accept \code{n}
as an argument.}

\item{add}{\code{TRUE} or \code{FALSE}. Should the plot be added to an
existing plot? Defaults to \code{FALSE}. Not applicable if
\code{use_ggplot = TRUE}. User could, \emph{e.g.}, use this option to
add this plot on top of \code{\link{map_dets}()}.}

\item{use_ggplot}{\code{TRUE} or \code{FALSE}. Should the plot be
made using \code{ggplot2}? Defaults to \code{FALSE} and uses base R
plotting.}
}
\description{
Plots the least cost paths between locations for all individuals
in an \code{lcp_list} object.
}
\details{
Blah blah
}
\examples{

#Load sample data
data(acoustic)

#Process detections and stations
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

#Calculate  COAs -- Note, using 6 hour COAs to reduce processing time
coas.6h <- coa_locs(proc.det, Delta_t = "6 hours")

#Initialize raster
r.cond <- init_raster(study_area = acoustic$study_area, res = 0.001)

#Create conductance raster
conduct <- raster_conduct(sa_rast = r.cond, impassable = acoustic$land)

#Create least-cost paths
lcps <- lc_path(det_locs = coas.6h, conductance = conduct)

#Plot
plot_lc_path(lcps)

#Plot on top of detections
  #Note set_par = FALSE in both functions
  #Consider manually setting graphical parameters for this plot

map_dets(proc_det = proc.det, sta_crs = 4326,
  base_layers = list(acoustic$study_area, acoustic$land),
  base_cols = list("gray30", "wheat"), base_borders = list(NA, "seagreen"),
  xlim=c(-64.636, -64.604), ylim=c(17.770, 17.795),
  leg_pos = "topleft", sta_col = "blue", sta_bg = "orange",
  set_par = FALSE)

plot_lc_path(lcps, set_par = FALSE, add = TRUE)

#Plot with ggplot
plot_lc_path(lcps, use_ggplot = TRUE) +
  theme_bw() +
  geom_sf(data = acoustic$land)
}
