% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_proc_inputs.R
\name{proc_dets}
\alias{proc_dets}
\alias{is.dets}
\title{Combine detections with stations}
\usage{
proc_dets(det, sta, crs = 4326)

is.dets(x)
}
\arguments{
\item{det}{A \code{data.frame} of detections. Format should match the
object \code{detections} from \link{acoustic} (\code{id}, \code{dt},
\code{rec_id}).}

\item{sta}{A \code{data.frame} of the stations where receivers were
located for a period of time. Format should match the object
\code{stations} from \link{acoustic} (\code{sta_id}, \code{rec_id},
\code{dt_dep}, \code{dt_ret}, \code{x}, \code{y}).}

\item{crs}{Coordinate Reference System to use for the detections. Passed
to \code{\link[sf:st_crs]{sf::st_crs}()} to set CRS for sf object.
Defaults to \code{4326}, longitude/latitude on the WGS84 spheroid.}

\item{x}{An object to check with \code{is.dets()}.}
}
\value{
Returns a \code{data.frame} of class \code{dets} of
georeferenced detections. Object is also of class \code{sf} and
contains column \code{geometry}.
}
\description{
This function combines detections with stations to place detections
in space.
}
\details{
Uses \code{\link[sqldf]{sqldf}} to combine \code{det} with
\code{sta} using \code{rec_id} where the date-time of the detection falls
within the date-time of the station deployment.

It specifically excutes the following SQL \code{SELECT} query:

\preformatted{
 SELECT id, sta.rec_id, sta_id, dt, lon, lat
 FROM det LEFT JOIN sta
 ON det.rec_id = sta.rec_id
     AND det.dt > sta.dt_dep
     AND det.dt < sta.dt_ret}

The left join keeps all rows from the detections
\code{data.frame}, inserting an \code{NA} if there is a
detection not within one of the time intervals for a given
receiver. The function subsequently removes any locations
with \code{NA} in the x- or y-coordinate with a warning.
}
\section{Functions}{
\itemize{
\item \code{is.dets}: Checks whether an object is of class \code{dets}.
}}

\examples{
#Load the example data set
data(acoustic)

#Process detections
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

}
\seealso{
\link{acoustic} for details on the formatting of the
\code{data.frame}s \code{det} and \code{sta}
}
