% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{raster_conduct}
\alias{raster_conduct}
\title{Create a conductance raster}
\usage{
raster_conduct(sa_rast, impassable, impassable_value = 100)
}
\arguments{
\item{sa_rast}{A \code{\link[raster:Raster-class]{RasterLayer}} object
on which to calculate conductance.}

\item{impassable}{An \code{\link[sf]{sfc_POLYGON}} object that contains
land or other impassable terrain.}

\item{impassable_value}{A \emph{resistance} value to assign to the
\code{impassable} cells of the raster. This should be a large number
to ensure the animal will not cross the impassable sections. The function
will convert it from a resistance to a conductance.}
}
\value{
Returns a \code{\link[raster:Raster-class]{RasterLayer}} object
of conductance values.
}
\description{
Creates a conductance raster for the study area for use in calculating
least cost paths between two locations.
}
\details{
This function takes a \code{\link[raster:Raster-class]{RasterLayer}}
as a template and an \code{\link[sf]{sfc_POLYGON}} and returns another
\code{RasterLayer} of conductance. This conductance raster is intended to
be passed downstream for calculation of least cost paths between animal
locations.
}
\examples{
#Load sample data
data(acoustic)

#Initialize raster
r.cond <- init_raster(study_area = acoustic$study_area, res = 0.001)

#Create conductance raster
conduct <- raster_conduct(sa_rast = r.cond, impassable = acoustic$land)

#Plot
plot(conductance)
plot(acoustic$study_area, add = TRUE)
plot(acoustic$land, add = TRUE, col = NA)

}
