% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_grids.R
\name{raster_overlap}
\alias{raster_overlap}
\title{Create an overlap raster}
\usage{
raster_overlap(..., background = NA)
}
\arguments{
\item{r}{A raster}
}
\description{
Takes multiple rasters of presence-only or presence-absence and
calculates the overlap.
}
\details{
This function sums the values of all the input rasters
and then replaces any 0 values with the background value.
}
