% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_grids.R
\name{raster_pres}
\alias{raster_pres}
\title{Create a presence raster}
\usage{
raster_pres(r)
}
\arguments{
\item{r}{A raster}
}
\description{
Converts a raster of counts to a raster of presence-only or
presence-absence
}
\details{
This function sets any values >= 1 to 1. If the background
values of the raster are NA, they remain NA in the output, allowing
the raster to be interpreted as presence-only. If the background
values of the raster are 0, then the result will be also contain 0s,
allowing the raster to be interpreted as presence-abscence.
}
