% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_grids.R
\name{rasterize_dets}
\alias{rasterize_dets}
\title{Rasterize acoustic detections}
\usage{
rasterize_dets(det_locs, r = init_raster(sta = det_locs))
}
\arguments{
\item{det_locs}{A \code{data.frame} of georeferenced acoustic detections.
Could be, \emph{e.g.}, the \code{data.frame} returned by
\code{\link{proc_dets}()} or the \code{data.frame} returned by
\code{\link{coa_locs}()}}

\item{r}{A blank raster to use as the template for rasterizing the
locations. Defaults to a call to \code{\link{init_rast}()} with
\code{det_locs} passed as the argument \code{sta}}
}
\value{
Returns an object of class \code{RasterLayer} with values equal
to the number of detections in each raster cell.
}
\description{
This function takes either detections at stations or centers of
activity (COAs) and creates a raster map showing number of detections
in each cell.
}
\details{
Calls the function \code{\link{raster::rasterize}()} to generate
the raster. Uses the function \code{'count'}
}
\examples{
#Load example data
data(acoustic)

#Process detections
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

#Default rasterize
r1 <- rasterize_dets(det_locs = proc.det)

#Plot
plot(r1)
plot(acoustic$land, add = TRUE, col="wheat")

#Rasterize with a smaller grid
r2 <- rasterize_dets(det_locs = proc.det,
                     r = init_raster(sta = proc.det, res = 0.001))

#Plot
plot(r2)
plot(acoustic$land, add = TRUE, col="wheat")

#Rasterize COAs
coas <- coa_locs(proc.det)
r3 <- rasterize_dets(det_locs = coas,
                     r = init_raster(sta = proc.det, res = 0.001))

#Plot
plot(r3)
plot(acoustic$land, add = TRUE, col="wheat")

}
