% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{regular_points}
\alias{regular_points}
\title{Generate regular points along a path}
\usage{
regular_points(paths, Delta_t)
}
\arguments{
\item{paths}{An object of class \code{*_paths}: either \code{str_paths}
or \code{lc_paths}}

\item{Delta_t}{The timestep to use for interpolation. The number of
interpolated points will be approximately (total time)/Delta_t. See
details below.}
}
\value{
Returns an object with S3 Class \code{reg_points}. Object is also
of class \code{sf}, \code{tbl_df}, \code{tbl}, and \code{data.frame}.
}
\description{
Takes a *_paths object and generates regularly-spaced
points along the path.
}
\details{
This function returns points that are regularly placed in
\emph{space}, even though the user provides a \emph{time} argument
(\code{Delta_t}). When the input locations are regularly spaced in time,
the output locations will be regularly spaced in both
\emph{space and time}. This will already be the case if the user has
decided to use COAs to represent the locations.

The number of points returned is reported in the output \code{data.frame}
in the column \code{$n_interp}. We determine this number by counting the
length of the sequence from the start time to the end time of the
track for each individual, with the \code{by} argument being
\code{Delta_t}. \emph{I.e.}:

\preformatted{
  n_interp = length(
               seq(
                 from = min(dt),
                 to = max(dt),
                 by = Delta_t))}

Therefore, \code{Delta_t} can be any value that can be accepted by
the \code{by} argument of \code{\link{seq.POSIXt}()}.
}
