% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_filter.R
\name{singleton_filter}
\alias{singleton_filter}
\title{Filters out singleton detections}
\usage{
singleton_filter(proc_det, time_horizon = "1 day")
}
\arguments{
\item{proc_det}{A \code{data.frame} of class \code{dets} as returned
by the function \code{\link{proc_dets}()}.}

\item{time_horizon}{The time during which a single detection will be
filtered. Can be any value that can be accepted by the \code{by} argument
of \code{\link{seq.POSIXt}()}.}
}
\description{
Filters out detections if only 1 detection occurred for an ID in
a user-defined time horizon.
}
\details{
False-positive detections in acoustic data can occur for reasons
related to signal collisions. Type A false detections are detections of
IDs that are not present in the study system, and here we assume the user
has already removed these detections prior to using this package. Type B
false detections occur when a signal is erroneously decoded as the signal
of a different tag that is in the study system -- these detections are
much harder to distinguish from true detections and are the focus of this
function.

This function is based on the assumption that it is possible for signal
collisions to result in a single false detection over a sufficiently long
time horizon, but that the probability of two false detections of the same
ID over that same time horizon is very low. Thus, the choice of time horizon
is very critical here. The period of time should be long enough that your
animal would plausibly use a receiver many times, but not so long that the
probability of two false detections of the same ID becomes unacceptably
high.
}
\examples{

#Load the example data set
data(acoustic)

#Process detections
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

#Apply the singleton filter with default time_horizon
det.filt <- singleton_filter(proc.det)

#Apply the singleton filter with 1 month time horizon
##Note: The user should see ?lubridate::add_with_rollback to understand
  #the behavior when adding months.
det.filt2 <- singleton_filter(proc.det, time_horizon = "1 month")

}
\seealso{
#' See Simpfendorfer \emph{et al.}
(\href{https://doi.org/10.1186/s40317-015-0094-z}{2015})
for details about false detections in acoustic telemetry.
}
