% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_filter.R
\name{spd_filter}
\alias{spd_filter}
\title{Filters out detections based on maximum speed over distance}
\usage{
spd_filter(proc_det, max_speed = units::set_units(10, "m/s"),
  min_dist = units::set_units(1, "km"))
}
\description{
DOCUMENT ME, PLEASE
}
