% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s5_interpolate_animate.R
\name{str_paths}
\alias{str_paths}
\title{Connect successive locations with lines}
\usage{
str_paths(det_locs)
}
\description{
Creates straight path line objects between each location for
any number of individuals.
}
