% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_proc_inputs.R
\name{zone_from_ll}
\alias{zone_from_ll}
\title{Find the UTM zone for a given location}
\usage{
zone_from_ll(lon, lat)
}
\arguments{
\item{lon}{A numeric vector of longitudes}

\item{lat}{A numeric vector of latitudes}
}
\value{
Returns an integer vector of UTM zones the same length as the
inputs \code{lon} and \code{lat}
}
\description{
This function returns the UTM zone for a given set of latitude/longitude.
}
\details{
This function takes a point and calculates the correct UTM zone
in which that point falls. It was developed based on answers to a
Stack Overflow post (see below).
}
\examples{
#UTM grid for Ft. Lauderdale, FL, USA
zone_from_ll(lon = -80.15, lat = 26.133333)

#UTM grid for Rome, Italy; Paris, France; London, UK
ll_df <- data.frame(city = c("Rome", "Paris", "London"),
                    lon = c(12.5, 2.3508, -0.1275),
                    lat = c(41.9, 48.8567, 51.507222))
zone_from_ll(lon = ll_df$lon, lat = ll_df$lat)

}
\references{
Based on the answers to the Stack Overflow post found here: \cr
\url{https://stackoverflow.com/questions/9186496/determining-utm-zone-to-convert-from-longitude-latitude}
}
