% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_proc_inputs.R
\name{proc_dets}
\alias{proc_dets}
\alias{is.dets}
\title{Combine detections with stations}
\usage{
proc_dets(det, sta, ...)

is.dets(x)
}
\arguments{
\item{det}{A \code{data.frame} of detections. Format should match the
object \strong{detections} from \link{acoustic} (`id`, `dt`, `rec_id`).}

\item{sta}{A \code{data.frame} of the stations where receivers were
located for a period of time. Format should match the object
\strong{stations} from \link{acoustic} (`sta_id`, `rec_id`, `dt_dep`,
`dt_ret`, `x`, `y`).}

\item{crs}{Coordinate Reference System to use for the detections. Passed
to \code{\link[sf:st_crs]{sf::st_crs}()} to set CRS for sf object.}

\item{...}{Additional arguments (not currently implemented)}
}
\value{
Returns a \code{data.frame} of class \code{dets} of
georeferenced detections. Object is also of class \code{sf} and
contains column \code{geometry}.
}
\description{
This function combines detections with stations to place detections
in space.

Convenience function that checks is object is of class \code{dets}.
}
\details{
Uses \code{\link[sqldf]{sqldf}} to combine \code{det} with
\code{sta} using \code{rec_id} where the date-time of the detection falls
within the date-time of the station deployment.

It specifically excutes the following SQL \code{SELECT} query:

\code{SELECT id, sta.rec_id, sta_id, dt, lon, lat} \cr
\code{FROM det LEFT JOIN sta}\cr
\code{ON det.rec_id = sta.rec_id}\cr
\code{AND det.dt > sta.dt_dep}\cr
\code{AND det.dt < sta.dt_ret}

The left join keeps all rows from the detections
\code{data.frame}, inserting an \code{NA} if there is a
detection not within one of the time intervals for a given
receiver. The function subsequently removes any locations
with \code{NA} in the x- or y-coordinate with a warning.
}
\section{Functions}{
\itemize{
\item \code{is.dets}: 
}}

\examples{
#Load the example data set
data(acoustic)

#Process detections
proc.det <- proc_dets(det = acoustic$detections, sta = acoustic$stations)

}
\seealso{
\link{acoustic} for details on the formatting of the
\code{data.frame}s \code{det} and \code{sta}
}
